Move Webauthn listener thread to WebauthnListener class
